# frontend/docs/bff-openapi.json
{
  "definitions": {
...
    "operationspb.CreateAirportRequest": {
      "properties": {
        "continent": {
          "type": "string"
        },
        "country_id": {
          "type": "string"
        },
        "elevation_ft": {
          "type": "number"
        },
        "fuel_price_multiplier": {
          "type": "number"
        },
        "gate_fee": {
          "type": "number"
        },
        "geog": {
          "type": "string"
        },
        "geom": {
          "type": "string"
        },
        "home_link": {
          "type": "string"
        },
        "iata_code": {
          "type": "string"
        },
        "icao_code": {
          "type": "string"
        },
        "intl_name": {
          "type": "string"
        },
        "local_name": {
          "type": "string"
        },
        "maintenance_point_price": {
          "type": "number"
        },
        "max_runway_length_m": {
          "type": "number"
        },
        "max_runway_uses_per_day": {
          "type": "number"
        },
        "municipality": {
          "type": "string"
        },
        "region_id": {
          "type": "string"
        },
        "runway_fee": {
          "type": "number"
        },
        "stand_fee": {
          "type": "number"
        },
        "timezone": {
          "type": "string"
        },
        "turnaround_point_price": {
          "type": "number"
        },
        "wikipedia_link": {
          "type": "string"
        },
        "works_at_night": {
          "type": "boolean"
        }
      },
      "type": "object"
    },
...
    "operationspb.IDResponse": {
      "properties": {
        "id": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "host": "localhost:4200",
  "info": {
    "contact": {
      "email": "lboris201@yandex.ru",
      "name": "Jeno7u",
      "url": "https://github.com/Jeno7u"
    },
    "title": "AirlineSim BFF API",
    "version": "1.0",
    "description": "Backend-for-frontend API. Based on backend OpenAPI and extended with BFF cache/filter query parameters."
  },
  "paths": {
...
    "/airport": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns",
        "parameters": [
          {
            "description": "Airport details",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/operationspb.CreateAirportRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Airport created successfully, id returned",
            "schema": {
              "$ref": "#/definitions/operationspb.IDResponse"
            }
          },
          "400": {
            "description": "1 - request validation error, 2 - country with such country_id dont exists, 3 - region with such region_id dont exists",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Airport with such ICAO/IATA already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create Airport (admin only)",
        "tags": [
          "Airport"
        ]
      }
    },
    "/airport/{id}": {
      "delete": {
        "description": "Returns deleted airport id",
        "parameters": [
          {
            "description": "Airport ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Airport deleted",
            "schema": {
              "$ref": "#/definitions/operationspb.IDResponse"
            }
          },
          "400": {
            "description": "1 - invalid id",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Airport not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Delete airport (admin only)",
        "tags": [
          "Airport"
        ]
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns patched airport id",
        "parameters": [
          {
            "description": "Airport ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "description": "Airport details",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/operationspb.ChangeAirportRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Airport patched successfully",
            "schema": {
              "$ref": "#/definitions/operationspb.IDResponse"
            }
          },
          "400": {
            "description": "1 - request validation error, 2 - country not found, 3 - region not found",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Airport not found"
          },
          "409": {
            "description": "Airport conflict"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Change airport (admin only)",
        "tags": [
          "Airport"
        ]
      }
    },
...
    "/countries": {
      "get": {
        "description": "Returns all countries\n\nBFF caches this list in memory, retries safe backend failures within BFF_BACKEND_MAX_SAFE_ATTEMPTS, and supports exact field filters plus q search across text fields.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Countries list",
            "schema": {
              "$ref": "#/definitions/operationspb.ListCountriesResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "List countries",
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "description": "Search query applied to all string fields.",
            "in": "query",
            "name": "q",
            "required": false,
            "type": "string"
          },
          {
            "description": "Force BFF to refresh its in-memory cache from backend.",
            "enum": [
              "true",
              "false"
            ],
            "in": "query",
            "name": "refresh",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by aircraft_tail_code.",
            "in": "query",
            "name": "aircraft_tail_code",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by corp_tax_rate.",
            "in": "query",
            "name": "corp_tax_rate",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by flythrough_permission_price.",
            "in": "query",
            "name": "flythrough_permission_price",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by id.",
            "in": "query",
            "name": "id",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by intl_name.",
            "in": "query",
            "name": "intl_name",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by iso.",
            "in": "query",
            "name": "iso",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by land_permission_price.",
            "in": "query",
            "name": "land_permission_price",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by local_name.",
            "in": "query",
            "name": "local_name",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by vat_rate.",
            "in": "query",
            "name": "vat_rate",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by wikipedia_link.",
            "in": "query",
            "name": "wikipedia_link",
            "required": false,
            "type": "string"
          }
        ]
      }
    },
    "/country": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns",
        "parameters": [
          {
            "description": "Country details",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/operationspb.CreateCountryRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Country created successfully, id returned",
            "schema": {
              "$ref": "#/definitions/operationspb.IDResponse"
            }
          },
          "400": {
            "description": "1 - request validation error",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Country with such ISO already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create country (admin only)",
        "tags": [
          "Country"
        ]
      }
    },
...
    "/region": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns",
        "parameters": [
          {
            "description": "Region details",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/operationspb.CreateRegionRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Region created successfully, id returned",
            "schema": {
              "$ref": "#/definitions/dto.IDResponse"
            }
          },
          "400": {
            "description": "1 - request validation error, 2 - country with such country_id dont exists",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Region with such local code already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create Region (admin only)",
        "tags": [
          "Region"
        ]
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "swagger": "2.0"
}